Rename to wind_cube_nacelle, drop nwtc
