+++ ingest/wind_cube_nacelle/config/pipeline_config_sta.yml
