+++ ingest/wind_cube_nacelle/config/storage_config.yml
