+++ ingest/wind_cube_nacelle/mapping.py
@@ -5,7 +5,6 @@ from utils import IngestSpec, expand
 from . import Pipeline
 
 
-
 # See https://regex101.com for information on setting up a regex pattern. Note that the
 # full filepath will be passed to the compiled regex pattern, so you can optionally
 # match the directory structure in addition to (or instead of) the file basename.
@@ -25,7 +24,7 @@ mapping: Dict["AnyStr@compile", IngestSpec] = {
         pipeline=Pipeline,
         pipeline_config=expand("config/pipeline_config_rtd.yml", __file__),
         storage_config=expand("config/storage_config.yml", __file__),
-        name="plot_wind_cube_nacelle_nwtc",
+        name="plot_wind_cube_nacelle",
     ),
     ## 10-min avg data
     # Mapping for Raw Data -> Ingest
@@ -40,6 +39,6 @@ mapping: Dict["AnyStr@compile", IngestSpec] = {
         pipeline=Pipeline,
         pipeline_config=expand("config/pipeline_config_rtd.yml", __file__),
         storage_config=expand("config/storage_config.yml", __file__),
-        name="plot_wind_cube_nacelle_nwtc",
-    ),
+        name="plot_wind_cube_nacelle",
+    ),wind_cube_nacelle
 }
