+++ ingest/wind_cube_nacelle/pipeline/pipeline.py
