+++ ingest/wind_cube_nacelle/pipeline/styling.mplstyle
