+++ ingest/wind_cube_nacelle/runner.py
@@ -1,4 +1,4 @@
-from ingest.wind_cube_nacelle_nwtc import Pipeline
+from ingest.wind_cube_nacelle import Pipeline
 from utils import expand, set_dev_env
