+++ ingest/wind_cube_nacelle/tests/data/expected/nwtc.wind_cube_nacelle_rtd.b0.20220112.181801.nc
