+++ ingest/wind_cube_nacelle/tests/data/expected/nwtc.wind_cube_nacelle_sta.b0.20220112.182000.nc
