+++ ingest/wind_cube_nacelle/tests/data/input/WIPO0200010-(wi020200010)_real_time_data_2022-01-12_19-00-00.csv
